// src/models/tablename.ts
enum TableName {
  admin = 'admin',
  adminSetting = 'admin_setting',
  allowanceAlbum = 'allowance_album',
  allowanceAlbumImage = 'allowance_album_image',
  allowanceChat = 'allowance_chat',
  allowanceHistory = 'allowance_history',
  allowanceMission = 'allowance_mission',
  allowanceSettlement = 'allowance_settlement',
  allowanceTask = 'allowance_task',
  appVersion = 'app_version',
  block = 'block',
  buddy = 'buddy',
  chatMessage = 'chat_message',
  chatRoom = 'chat_room',
  chatUser = 'chat_user',
  codeBank = 'code_bank',
  payment = 'payment',
  profile = 'profile',
  report = 'report',
  user = 'user',
  userSetting = 'user_setting'
  }
  
  export default TableName